# Keep dynamic batches about the same shape