@stephenroller 2020-08-02
Putting this back as a draft. It's turned out to be deviously trickly to get right. I think the Teacher rewrite may need to come first.