@klshuster 2020-09-08
@stephenroller is this something we should look back into? from your experiments it seemed to have worked pretty well in terms of speedup